+++ CHANGELOG.md
@@ -512,3 +512,5 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 
 #### dev ####
 * Allow Dynamic Variables to show in smith view upgraded (kiooeht)
+* Maybe make card description patches work for Chinese:
+  * Shrink long descriptions (kiooeht)